pbuilder: build in $WORKSPACE or tmpfs if available
Come to find out unless --buildplace is defined, pbuilder builds in /var/cache/pbuilder/build.

The braggi and adami builders have separate faster and larger SSDs mounted at /home/jenkins-build but pbuilder hasn't been using them.

Additionally, it is possible to build in memory.  The braggi and adami have plenty of space available to build in it so let's use it.

Signed-off-by: David Galloway <dgallowa@redhat.com>